requirements.txt
django==5.1.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
pymysql==1.1.1
django-shortuuidfield==0.1.3
pyjwt==2.9.0
